change bloom filter strategy

+++ strategy.go
@@ -18,11 +18,11 @@ type MURMUR128MITZ64 struct {
 func (strategy *MURMUR128MITZ64) Indexes(data []byte, m uint, k uint) []uint {
 	indexes := make([]uint, k)
 	h1, h2 := murmur3.Sum128(data)
-	combined := uint64(h1)
+	combined := int64(h1)
 
 	for i := uint(0); i < k; i++ {
-		indexes[i] = uint(combined&uint64(0x7fffffffffffffff)) % uint(m)
-		combined += uint64(h2)
+		indexes[i] = uint((combined & int64(0x7fffffffffffffff)) % int64(m))
+		combined += int64(h2)
 	}
 	return indexes
 }
